[MAC support] Move around platform classes to increase code reuse
