+++ src/Microsoft.Identity.Client/Cache/Keys/MsalAccessTokenCacheKey.cs
@@ -28,6 +28,7 @@
 using System;
 using System.Text;
 using Microsoft.Identity.Client.Core;
+using Microsoft.Identity.Client.PlatformsCommon.Factories;
 
 namespace Microsoft.Identity.Client.Cache
 {
