+++ src/Microsoft.Identity.Client/CacheV2/Impl/FileSystemCredentialPathManager.cs
@@ -31,6 +31,7 @@ using Microsoft.Identity.Client.CacheV2.Impl.Utils;
 using Microsoft.Identity.Client.CacheV2.Schema;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client;
+using Microsoft.Identity.Client.PlatformsCommon.Factories;
 
 namespace Microsoft.Identity.Client.CacheV2.Impl
 {
