+++ src/Microsoft.Identity.Client/Features/ConfidentialClient/ClientCredential.cs
@@ -29,7 +29,7 @@ using System;
 
 namespace Microsoft.Identity.Client
 {
-#if !ANDROID_BUILDTIME && !iOS_BUILDTIME && !WINDOWS_APP_BUILDTIME // Hide confidential client on mobile platforms
+#if !ANDROID_BUILDTIME && !iOS_BUILDTIME && !WINDOWS_APP_BUILDTIME && !MAC_BUILDTIME // Hide confidential client on mobile platforms
 
     /// <summary>
     /// Meant to be used in confidential client applications, an instance of <c>ClientCredential</c> is passed
