+++ src/Microsoft.Identity.Client/Features/ConfidentialClient/IConfidentialClientApplication.cs
@@ -31,7 +31,7 @@ using System.Threading.Tasks;
 
 namespace Microsoft.Identity.Client
 {
-#if !ANDROID_BUILDTIME && !iOS_BUILDTIME && !WINDOWS_APP_BUILDTIME // Hide confidential client on mobile platforms
+#if !ANDROID_BUILDTIME && !iOS_BUILDTIME && !WINDOWS_APP_BUILDTIME && !MAC_BUILDTIME // Hide confidential client on mobile platforms
 
     /// <summary>
     /// Component to be used with confidential client applications like Web Apps/API.
