+++ src/Microsoft.Identity.Client/Features/PublicClientWithTokenCache/TokenCacheAccessor.cs
