+++ src/Microsoft.Identity.Client/Internal/Requests/AuthenticationRequestParameters.cs
@@ -55,7 +55,7 @@ namespace Microsoft.Identity.Client.Internal.Requests
         public string SliceParameters { get; set; }
         public bool SendCertificate { get; set; }
 
-#if !ANDROID_BUILDTIME && !iOS_BUILDTIME && !WINDOWS_APP_BUILDTIME // Hide confidential client on mobile platforms
+#if !ANDROID_BUILDTIME && !iOS_BUILDTIME && !WINDOWS_APP_BUILDTIME && !MAC_BUILDTIME // Hide confidential client on mobile platforms
         public ClientCredential ClientCredential { get; set; }
 #endif
