+++ src/Microsoft.Identity.Client/Platforms/Android/AndroidPlatformLogger.cs
@@ -27,6 +27,7 @@
 
 using Android.Util;
 using Microsoft.Identity.Client.Core;
+using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 
 namespace Microsoft.Identity.Client.Platforms.Android
 {
