+++ src/Microsoft.Identity.Client/Platforms/netcore/NetCorePlatformProxy.cs
@@ -30,7 +30,8 @@ using System.Reflection;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client.Cache;
 using Microsoft.Identity.Client.Core;
-using Microsoft.Identity.Client.Internal;
+using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
+using Microsoft.Identity.Client.PlatformsCommon.Shared;
 using Microsoft.Identity.Client.UI;
 
 namespace Microsoft.Identity.Client.Platforms.netcore
@@ -40,7 +41,8 @@ namespace Microsoft.Identity.Client.Platforms.netcore
     /// </summary>
     internal class NetCorePlatformProxy : IPlatformProxy
     {
-        private readonly Lazy<IPlatformLogger> _platformLogger = new Lazy<IPlatformLogger>(() => new EventSourcePlatformLogger());
+        private readonly Lazy<IPlatformLogger> _platformLogger = new Lazy<IPlatformLogger>(
+            () => new EventSourcePlatformLogger());
         private IWebUIFactory _overloadWebUiFactory;
 
         /// <summary>
@@ -132,7 +134,7 @@ namespace Microsoft.Identity.Client.Platforms.netcore
 
         public ILegacyCachePersistence CreateLegacyCachePersistence()
         {
-            return new NetCoreLegacyCachePersistence();
+            return new InMemoryLegacyCachePersistance();
         }
 
         public ITokenCacheAccessor CreateTokenCacheAccessor()
