+++ src/Microsoft.Identity.Client/Platforms/netstandard13/NetStandard13CryptographyManager.cs
@@ -30,6 +30,7 @@ using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using Microsoft.Identity.Client.Core;
+using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 using Microsoft.Identity.Client.Utils;
 
 namespace Microsoft.Identity.Client.Platforms.netstandard13
