+++ src/Microsoft.Identity.Client/Platforms/uap/UapLegacyCachePersistance.cs
@@ -31,6 +31,7 @@ using Windows.Storage;
 using Microsoft.Identity.Client.Cache;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Internal;
+using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 
 namespace Microsoft.Identity.Client.Platforms.uap
 {
