+++ src/Microsoft.Identity.Client/Platforms/uap/UapTokenCacheAccessor.cs
@@ -29,6 +29,7 @@ using Windows.Storage;
 using Microsoft.Identity.Client.Cache;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Utils;
+using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 
 namespace Microsoft.Identity.Client.Platforms.uap
 {
