+++ src/Microsoft.Identity.Client/PlatformsCommon/Interfaces/ICryptographyManager.cs
@@ -1,20 +1,20 @@
-﻿//----------------------------------------------------------------------
-//
+﻿// ------------------------------------------------------------------------------
+// 
 // Copyright (c) Microsoft Corporation.
 // All rights reserved.
-//
+// 
 // This code is licensed under the MIT License.
-//
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files(the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions :
-//
+// 
 // The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE
@@ -22,38 +22,23 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-//
-//------------------------------------------------------------------------------
+// 
+// ------------------------------------------------------------------------------
 
-using System.Diagnostics.Tracing;
+using System.Security.Cryptography.X509Certificates;
 
-namespace Microsoft.Identity.Client.Internal.EventsSource
+namespace Microsoft.Identity.Client.PlatformsCommon.Interfaces
 {
-    [EventSource(Name = "Microsoft.Identity.Client")]
-    internal class MsalEventSource : EventSource
+    internal interface ICryptographyManager
     {
-        [Event(1, Level = EventLevel.Verbose)]
-        internal void Verbose(string message)
-        {
-            WriteEvent(1, message);
-        }
-
-        [Event(2, Level = EventLevel.Informational)]
-        internal void Information(string message)
-        {
-            WriteEvent(2, message);
-        }
-
-        [Event(3, Level = EventLevel.Warning)]
-        internal void Warning(string message)
-        {
-            WriteEvent(3, message);
-        }
-
-        [Event(4, Level = EventLevel.Error)]
-        internal void Error(string message)
-        {
-            WriteEvent(4, message);
-        }
+        string CreateBase64UrlEncodedSha256Hash(string input);
+        string GenerateCodeVerifier();
+        string CreateSha256Hash(string input);
+        byte[] CreateSha256HashBytes(string input);
+        string Encrypt(string message);
+        string Decrypt(string encryptedMessage);
+        byte[] Encrypt(byte[] message);
+        byte[] Decrypt(byte[] encryptedMessage);
+        byte[] SignWithCertificate(string message, X509Certificate2 certificate);
     }
 }
