+++ src/Microsoft.Identity.Client/PlatformsCommon/Interfaces/IPlatformLogger.cs
@@ -1,46 +1,37 @@
-﻿//----------------------------------------------------------------------
-//
+﻿// ------------------------------------------------------------------------------
+// 
 // Copyright (c) Microsoft Corporation.
 // All rights reserved.
-//
+// 
 // This code is licensed under the MIT License.
-//
+// 
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files(the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions :
-//
+// 
 // The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
-//
+// 
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-//
-//------------------------------------------------------------------------------
+// 
+// ------------------------------------------------------------------------------
 
-using Microsoft.Identity.Client.Cache;
-
-namespace Microsoft.Identity.Client.Platforms.netcore
+namespace Microsoft.Identity.Client.PlatformsCommon.Interfaces
 {
-    internal class NetCoreLegacyCachePersistence : ILegacyCachePersistence
+    internal interface IPlatformLogger
     {
-        private byte[] data;
-
-        byte[] ILegacyCachePersistence.LoadCache()
-        {
-            return data;
-        }
-
-        void ILegacyCachePersistence.WriteCache(byte[] serializedCache)
-        {
-            data = serializedCache;
-        }
+        void Error(string message);
+        void Warning(string message);
+        void Verbose(string message);
+        void Information(string message);
     }
 }
