+++ src/Microsoft.Identity.Client/PlatformsCommon/Interfaces/IPlatformProxy.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client.Cache;
+using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.UI;
 
-namespace Microsoft.Identity.Client.Core
+namespace Microsoft.Identity.Client.PlatformsCommon.Interfaces
 {
     /// <summary>
     /// Common operations for extracting platform / operating system specifics
