+++ src/Microsoft.Identity.Client/PlatformsCommon/Shared/ConsolePlatformLogger.cs
@@ -1,20 +1,20 @@
-﻿// ------------------------------------------------------------------------------
-// 
+﻿//------------------------------------------------------------------------------
+//
 // Copyright (c) Microsoft Corporation.
 // All rights reserved.
-// 
+//
 // This code is licensed under the MIT License.
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files(the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions :
-// 
+//
 // The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE
@@ -22,16 +22,37 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-// 
-// ------------------------------------------------------------------------------
+//
+//------------------------------------------------------------------------------
 
-namespace Microsoft.Identity.Client.Core
+
+using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
+using System;
+
+namespace Microsoft.Identity.Client.PlatformsCommon.Shared
 {
-    internal interface IPlatformLogger
+    #if !WINDOWS_APP
+    class ConsolePlatformLogger : IPlatformLogger
     {
-        void Error(string message);
-        void Warning(string message);
-        void Verbose(string message);
-        void Information(string message);
+        public void Error(string message)
+        {
+            Console.WriteLine(message);
+        }
+
+        public void Warning(string message)
+        {
+            Console.WriteLine(message);
+        }
+
+        public void Verbose(string message)
+        {
+            Console.WriteLine(message);
+        }
+
+        public void Information(string message)
+        {
+            Console.WriteLine(message);
+        }
     }
+    #endif
 }
