+++ src/Microsoft.Identity.Client/TelemetryCore/DefaultEvent.cs
@@ -28,6 +28,7 @@
 using System.Collections.Generic;
 using System.Globalization;
 using Microsoft.Identity.Client.Core;
+using Microsoft.Identity.Client.PlatformsCommon.Factories;
 
 namespace Microsoft.Identity.Client.TelemetryCore
 {
