+++ src/Microsoft.Identity.Client/TelemetryCore/Event.cs
@@ -31,6 +31,7 @@ using Microsoft.Identity.Client.Instance;
 using System.Globalization;
 using Microsoft.Identity.Client.Core;
 using Microsoft.Identity.Client.Utils;
+using Microsoft.Identity.Client.PlatformsCommon.Factories;
 
 namespace Microsoft.Identity.Client.TelemetryCore
 {
